Tests work on Windows 2019. Removing Travis.
I found that the default user for Windows 2019 on Azure Pipelines is
"VssAdministrator\u{0}", so I hardcoded the test case to expect that.
Since the macOS test case also needs to pass (since we are studying the
build system at the moment, and not yet focusing on implementation
details), I also hardcoded the macOS implementation of 'getusername' to
return the string "VssAdministrator\u{0}".